feat(admin): implement admin management and role-based conflict logic
- Add Admin registration, restoration, and role upgrade logic.
- Implement role conflict check to prevent concurrent Admin/Driver roles.
- Integrate MapStruct for efficient DTO mapping and partial updates.
- Add specialized operations for salary increments and role updates.

+++ src/main/java/com/nexus/NexusShip/model/Driver.java
@@ -45,7 +45,7 @@ public class Driver extends User{
     public Driver(){}
 
     public Driver(String firstName, String lastName, Gender gender, String nationalId, String email, String password,
-                  String phoneNumber, String licenseNumber, double salary) {
+                  String phoneNumber, String licenseNumber, BigDecimal salary) {
         super(firstName, lastName, gender, nationalId, email, password, phoneNumber);
 
         this.licenseNumber = licenseNumber;
